move out uint32
