+++ rust/blockstore/src/arrow/block/delta/mod.rs
@@ -1,6 +1,7 @@
 mod delta;
 mod storage;
 pub(super) mod string;
+pub(super) mod uint32;
 
 pub use delta::*;
 pub use storage::*;
